requirements.txt

langchain>=0.2.3
langchain-community>=0.2.4
python-dotenv>=1.0.1
openai>=1.38.0
